[QFEED-81] Refactor: 조회쿼리가 2번 나가는 문제 해결
- @Transaction의 위치를 서비스로 이동
